requirements.txt
stomp.py
pydantic
pydantic-settings